==> revision2/results_from_latest_bigsnpr_package/lassosum2_most_recent_version/Simulation/2_prs_computing.R <==

rm(list=ls())

library(bigsnpr)
library(bigreadr)
library(bigparallelr)

parameters <- expand.grid(race = c('AFR','AMR','EAS','EUR','SAS'),
                          setting = 1:5, rho = 1:3, size = 1:4, rep = 1:10)

args <- commandArgs(T)
for(i in 1:length(args)){ eval(parse(text=args[[i]])) }

race = as.character(parameters[as.numeric(temp1),1])
setting = as.integer(parameters[as.numeric(temp1),2])
rho = as.integer(parameters[as.numeric(temp1),3])
size = as.integer(parameters[as.numeric(temp1),4])
rep = as.integer(parameters[as.numeric(temp1),5])


print(setting)
print(rho)
print(size)
print(rep)

print(race)

NCORES <-  1

setwd(paste0("/dcs04/nilanjan/data/jzhang2/MEPRS/revision2/update_bigsnpr_results/Simulation/Setting_",setting,"/",race,"/rep_",rep))

dir.create(paste0("./lassosum2/prs/tuning/rho_",rho,"_size_",size), recursive=T)
dir.create(paste0("./lassosum2/prs/validation/rho_",rho,"_size_",size), recursive=T)

nsnps <- integer()
for (chr in 1:22){
  #prs.file <- fread2(paste0("./lassosum2/coef/rho_",rho,"_size_",size,"_lassosum2-chr",chr,".txt"))
  nsnps[chr] <- 122
}

################################################
################################################

## tuning

for (race1 in c('AFR','AMR','EAS','EUR','SAS')){

  print(paste0(race1))

  if(race=="EUR" & race1!="EUR"){
    next
  }

  for (chr in 1:22){

    print(paste0("tuning-",chr))

    system(paste0("/dcs04/nilanjan/data/jzhang2/TOOLS/plink/plink2 --threads ",NCORES,
                  " --bfile /dcs04/nilanjan/data/jzhang2/MEPRS/Simulation/geno/mega_matched_summary_stat/", race,"/tv_chr",chr,
                  " --keep /dcs04/nilanjan/data/jzhang2/MEPRS/Simulation/geno/mega_matched_summary_stat/", race,"/test.id.txt",
                  " --score ../../",race1,"/rep_",rep,"/lassosum2/coef/rho_",rho,"_size_",ifelse(race1=='EUR',4,size),"_lassosum2-chr",chr,".txt",
                  " cols=+scoresums,-scoreavgs --score-col-nums 3-", nsnps[chr],
                  " --out ./lassosum2/prs/tuning/rho_",rho,"_size_",size,"/",race,"-",race1,"_chr", chr,"_tuning_score"))

    score <- fread2(paste0("./lassosum2/prs/tuning/rho_",rho,"_size_",size,"/",race,"-",race1,"_chr", chr,"_tuning_score.sscore"))

    if(chr==1){
      SCORE_all <- score[, -1:-4,drop=F]
      SCORE_id <- score[,1:2,drop=F]
    }else{
      SCORE_all <- SCORE_all + score[, -1:-4,drop=F]
    }
  }
  SCORE <- SCORE_all
  save(SCORE, SCORE_id, file = paste0("./lassosum2/prs/tuning/rho_",rho,"_size_",size,"/",race,"-",race1,"_tuning_score.RData"))
  rm(list=c("SCORE","SCORE_id","score","SCORE_all"))
}

################################################
################################################

## validation

dir.create("./lassosum2/prs/validation")

for (race1 in c('AFR','AMR','EAS','EUR','SAS')){

  print(paste0(race1))

  if(race=="EUR" & race1!="EUR"){
    next
  }

  for (chr in 1:22){

    print(paste0("validation-",chr))

    system(paste0("/dcs04/nilanjan/data/jzhang2/TOOLS/plink/plink2 --threads ",NCORES,
                  " --bfile /dcs04/nilanjan/data/jzhang2/MEPRS/Simulation/geno/mega_matched_summary_stat/", race,"/tv_chr",chr,
                  " --keep /dcs04/nilanjan/data/jzhang2/MEPRS/Simulation/geno/mega_matched_summary_stat/", race,"/validation.id.txt",
                  " --score ../../",race1,"/rep_",rep,"/lassosum2/coef/rho_",rho,"_size_",ifelse(race1=='EUR',4,size),"_lassosum2-chr",chr,".txt",
                  " cols=+scoresums,-scoreavgs --score-col-nums 3-",nsnps[chr],
                  " --out ./lassosum2/prs/validation/rho_",rho,"_size_",size,"/",race,"-",race1,"_chr", chr,"_validation_score"))

    score <- fread2(paste0("./lassosum2/prs/validation/rho_",rho,"_size_",size,"/",race,"-",race1,"_chr", chr,"_validation_score.sscore"))
    if(chr==1){
      SCORE_all <- score[, -1:-4,drop=F]
      SCORE_id <- score[,1:2,drop=F]
    }else{
      SCORE_all <- SCORE_all + score[, -1:-4,drop=F]
    }
  }
  SCORE <- SCORE_all
  save(SCORE, SCORE_id, file = paste0("./lassosum2/prs/validation/rho_",rho,"_size_",size,"/",race,"-",race1,"_validation_score.RData"))
  rm(list=c("SCORE","SCORE_id","score","SCORE_all"))
}
